% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{add_arg}
\alias{add_arg}
\title{Add a new argument to a call}
\usage{
add_arg(expr, value, name = NULL, position = NULL)
}
\arguments{
\item{expr}{A call or list of calls.}

\item{value}{The value to add as an argument.}

\item{name}{Optional name for the new argument.}

\item{position}{Optional position to insert the argument (1 = first, etc.).}
}
\value{
The modified call (or list of calls).
}
\description{
Appends or inserts a new argument into a call expression.
}
