% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{append_meta_to_last}
\alias{append_meta_to_last}
\title{Append metadata to the most recent captured call}
\usage{
append_meta_to_last(call_list, meta)
}
\arguments{
\item{call_list}{List of call objects of class `callobj`}

\item{meta}{A named list of metadata}
}
\value{
Updated list with metadata applied to last item
}
\description{
Append metadata to the most recent captured call
}
\keyword{internal}
