% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{apply_inline_meta}
\alias{apply_inline_meta}
\title{Apply inline `meta()` calls to call list}
\usage{
apply_inline_meta(calls)
}
\arguments{
\item{calls}{A list of captured calls with metadata}
}
\value{
A new list with inline metadata applied
}
\description{
Apply inline `meta()` calls to call list
}
\keyword{internal}
