% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomize.R
\name{atomize_expr}
\alias{atomize_expr}
\title{Fully atomize an expression}
\usage{
atomize_expr(expr, prefix = "tmp", depth = Inf)
}
\arguments{
\item{expr}{A call object.}

\item{prefix}{Prefix for naming temporary variables.}

\item{depth}{Maximum recursion depth (default: Inf).}
}
\value{
A list of calls: assignments followed by the final call.
}
\description{
Breaks a call into a sequence of sub-expressions, assigning intermediate results
to temporary variables, up to a given depth.
}
