% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomize.R
\name{atomize_expr_with_counter}
\alias{atomize_expr_with_counter}
\title{Fully atomize an expression with an external counter and depth control}
\usage{
atomize_expr_with_counter(expr, prefix = "tmp", counter = 1, depth = Inf)
}
\arguments{
\item{expr}{A call object.}

\item{prefix}{Prefix for naming temporary variables.}

\item{counter}{Initial counter value.}

\item{depth}{Maximum recursion depth (default: Inf).}
}
\value{
A list containing `expressions` and updated `counter`.
}
\description{
Like `atomize_expr()` but allows a depth limit and initial counter to be passed in.
}
