% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomize.R
\name{atomize_selective_capture}
\alias{atomize_selective_capture}
\title{Selectively atomize parts of a code_capture object}
\usage{
atomize_selective_capture(capture, fn_names, prefix = "tmp", depth = Inf)
}
\arguments{
\item{capture}{A `code_capture` object.}

\item{fn_names}{Character vector of functions to atomize.}

\item{prefix}{Prefix for temporary variable names.}

\item{depth}{Maximum recursion depth (default: Inf).}
}
\value{
A new `code_capture` object.
}
\description{
Decomposes specified function calls into separate expressions across a `code_capture`,
respecting a recursion depth.
}
