% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{capture}
\alias{capture}
\title{Start a capture session or capture code}
\usage{
capture(x, script = NULL)
}
\arguments{
\item{x}{Optional: A code block (`{}`)}

\item{script}{Optional: script file path.}
}
\value{
A `code_capture` object, or starts a capture session.
}
\description{
Depending on the input:
- If no input, starts a session capture (top-level expressions).
- If given a `{}` block, captures it immediately.
- If given a script file path, captures and parses it.
}
