% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{capture_block}
\alias{capture_block}
\title{Capture a code block}
\usage{
capture_block(expr, realize = FALSE, evaluate = FALSE, envir = parent.frame())
}
\arguments{
\item{expr}{A `{}` block}

\item{realize}{Realize arguments (default FALSE)}

\item{evaluate}{Evaluate immediately (default FALSE)}

\item{envir}{Evaluation environment}
}
\value{
A `code_capture` object
}
\description{
Capture a code block
}
