% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{capture_end}
\alias{capture_end}
\title{End an active capture session}
\usage{
capture_end()
}
\value{
A `code_capture` object.
}
\description{
Processes the collected expressions and attaches metadata if present.
}
