% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{capture_object}
\alias{capture_object}
\title{Create a capture_object}
\usage{
capture_object(exprs = list())
}
\arguments{
\item{exprs}{A list of expressions (default: empty list).}
}
\value{
An object of class `capture_object`.
}
\description{
Defines a lightweight container for storing code expressions
with a custom class.
}
