% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{capture_script}
\alias{capture_script}
\title{Capture code from a script}
\usage{
capture_script(path, encoding = "UTF-8")
}
\arguments{
\item{path}{script file path.}

\item{encoding}{file encoding of script}
}
\value{
A `code_capture` object
}
\description{
Capture code from a script
}
