% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{$.callobj}
\alias{$.callobj}
\title{Extract an argument by name (shorthand)}
\usage{
\method{$}{callobj}(x, name)
}
\arguments{
\item{x}{A call object.}

\item{name}{Name of the argument.}
}
\description{
Extract an argument by name (shorthand)
}
