% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{change_arg}
\alias{change_arg}
\title{Change an existing argument in a call}
\usage{
change_arg(expr, name, new_value)
}
\arguments{
\item{expr}{A expression or callobj.}

\item{name}{The argument name to change.}

\item{new_value}{The new value to assign.}
}
\value{
The modified call (or list of calls).
}
\description{
Changes the value of an existing argument in a call. Throws an error if the argument does not exist.
}
