% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{change_func}
\alias{change_func}
\title{Change the function in a call}
\usage{
change_func(expr, new_func)
}
\arguments{
\item{expr}{A expression or callobj}

\item{new_func}{The new function name (symbol or character).}
}
\value{
The modified call (or list of calls).
}
\description{
Replaces the function name of a call while keeping its arguments.
}
