% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{compress_redundant_versions}
\alias{compress_redundant_versions}
\title{Compress redundant immutable variable versions}
\usage{
compress_redundant_versions(capture)
}
\arguments{
\item{capture}{A `code_capture` object.}
}
\value{
A new `code_capture` object with compressed variable names.
}
\description{
If a reassigned variable version is only used once to create another,
rewrites later version to use the earlier name and removes redundancy.
}
