% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose_capture}
\alias{diagnose_capture}
\title{Run diagnostics on a capture object}
\usage{
diagnose_capture(capture)
}
\arguments{
\item{capture}{A `code_capture` object.}
}
\value{
An object of class `"code_diagnosis"`.
}
\description{
Returns a structured list of diagnostic results indicating
stochasticity, side effects, non-standard evaluation, and more.
}
