% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose_capture_summary_fast}
\alias{diagnose_capture_summary_fast}
\title{Run diagnostics and return logical summary}
\usage{
diagnose_capture_summary_fast(capture)
}
\arguments{
\item{capture}{A `code_capture` object.}
}
\value{
A named list of diagnostics (TRUE/FALSE or vector).
}
\description{
Fast, non-evaluative diagnostics on a `code_capture` object.
}
\keyword{internal}
