% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{-.callobj}
\alias{-.callobj}
\title{Remove arguments from a call}
\usage{
\method{-}{callobj}(e1, e2)
}
\arguments{
\item{e1}{A call object.}

\item{e2}{A character vector (names) or numeric vector (positions).}
}
\description{
Remove arguments from a call
}
