% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{duplicate_line}
\alias{duplicate_line}
\title{Duplicate an expression in a code_capture object}
\usage{
duplicate_line(capture, i)
}
\arguments{
\item{capture}{A `code_capture` object.}

\item{i}{Integer index of the expression to duplicate.}
}
\value{
A modified `code_capture` object with duplicated expression.
}
\description{
Inserts a duplicate of the expression at index `i` directly after the original.
}
