% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{extract_named_meta}
\alias{extract_named_meta}
\title{Extract named metadata from a `meta()` call}
\usage{
extract_named_meta(expr)
}
\arguments{
\item{expr}{A `meta(...)` call expression}
}
\value{
A named list of metadata
}
\description{
Extract named metadata from a `meta()` call
}
\keyword{internal}
