% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_by_predicate}
\alias{filter_by_predicate}
\title{Filter expressions using a custom predicate}
\usage{
filter_by_predicate(capture, predicate)
}
\arguments{
\item{capture}{A `code_capture` object.}

\item{predicate}{A function taking a call and returning TRUE or FALSE.}
}
\value{
A filtered `code_capture` object.
}
\description{
Selects expressions for which the given predicate function returns TRUE.
}
