% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_all_arguments}
\alias{get_all_arguments}
\title{Get all arguments from all expressions in a capture}
\usage{
get_all_arguments(capture)
}
\arguments{
\item{capture}{A `code_capture` object.}
}
\value{
A list of argument lists.
}
\description{
Extracts arguments for each call in a `code_capture` object.
}
