% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{get_argument_names}
\alias{get_argument_names}
\title{Get the names of arguments in a call}
\usage{
get_argument_names(expr)
}
\arguments{
\item{expr}{A call object.}
}
\value{
A character vector of argument names.
}
\description{
Returns the argument names (or empty strings for unnamed arguments).
}
