% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_arguments}
\alias{get_arguments}
\title{Get the arguments from a call}
\usage{
get_arguments(expr)
}
\arguments{
\item{expr}{A call object.}
}
\value{
A list of arguments, or NULL if not a call.
}
\description{
Returns a list of arguments from a call expression.
}
