% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_expr_text}
\alias{get_expr_text}
\title{Get the deparsed text of expressions in a capture}
\usage{
get_expr_text(capture, collapse = "\\n")
}
\arguments{
\item{capture}{A `code_capture` object.}

\item{collapse}{String used to collapse multi-line expressions.}
}
\value{
A character vector with one element per expression.
}
\description{
Converts all expressions in a `code_capture` object to character strings.
}
