% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_expressions}
\alias{get_expressions}
\title{Access expressions from a capture object}
\usage{
get_expressions(capture)
}
\arguments{
\item{capture}{A `code_capture` object}
}
\value{
A list of expressions
}
\description{
Access expressions from a capture object
}
