% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{get_function}
\alias{get_function}
\title{Get the function called in an expression}
\usage{
get_function(expr)
}
\arguments{
\item{expr}{A call object.}
}
\value{
The function name or call (symbol or call).
}
\description{
Extracts the function from a call expression.
}
