% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inout.R
\name{get_inputs}
\alias{get_inputs}
\title{Identify inputs used before they are assigned}
\usage{
get_inputs(capture)
}
\arguments{
\item{capture}{A code_capture object.}
}
\value{
A named list mapping each input symbol to a list with its first usage index
        and whether it was used before definition.
}
\description{
Examines a code_capture object and returns inputs that are used before being defined.
}
