% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_metadata}
\alias{get_metadata}
\title{Access metadata from a capture object}
\usage{
get_metadata(capture)
}
\arguments{
\item{capture}{A `code_capture` object}
}
\value{
A list of metadata
}
\description{
Access metadata from a capture object
}
