% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{get_operator}
\alias{get_operator}
\title{Get the operator of a call}
\usage{
get_operator(expr)
}
\arguments{
\item{expr}{A call or callobj.}
}
\value{
Character name of operator, or NULL.
}
\description{
Returns the operator as character, or NULL if not an operator.
}
