% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{get_rhs}
\alias{get_rhs}
\title{Get the right-hand side of an assignment}
\usage{
get_rhs(expr)
}
\arguments{
\item{expr}{A call object.}
}
\value{
The RHS expression.
}
\description{
Extracts the value or expression assigned in an assignment call.
}
