% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{has_arg}
\alias{has_arg}
\title{Check if a call has a specific argument}
\usage{
has_arg(expr, name)
}
\arguments{
\item{expr}{A call object.}

\item{name}{The argument name to check for.}
}
\value{
Logical TRUE/FALSE.
}
\description{
Tests whether a named argument exists in a function call.
}
