% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{has_operator}
\alias{has_operator}
\title{Check if a call has an operator as function}
\usage{
has_operator(expr)
}
\arguments{
\item{expr}{A call or callobj.}
}
\value{
TRUE if the function name is a base operator.
}
\description{
Check if a call has an operator as function
}
