% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{is_assignment}
\alias{is_assignment}
\title{Check if an expression is an assignment}
\usage{
is_assignment(expr)
}
\arguments{
\item{expr}{A call or list of calls.}
}
\value{
Logical TRUE/FALSE (or logical vector if input is a list).
}
\description{
Detects whether the top-level call is an assignment (`<-` or `=`).
}
