% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_call_or_list}
\alias{is_call_or_list}
\title{Check if an object is a call or a list of calls}
\usage{
is_call_or_list(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
Logical TRUE/FALSE.
}
\description{
Determines whether the input is a single function call or a list entirely composed of calls.
}
