% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{is_compound}
\alias{is_compound}
\title{Check if an expression is a compound (nested) call}
\usage{
is_compound(expr)
}
\arguments{
\item{expr}{A call or callobj.}
}
\value{
Logical TRUE/FALSE.
}
\description{
Returns TRUE if any arguments of the call are themselves calls.
}
