% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{is_function}
\alias{is_function}
\title{Check if an expression is a function call}
\usage{
is_function(expr)
}
\arguments{
\item{expr}{An expression.}
}
\value{
TRUE if expr is a call, FALSE otherwise.
}
\description{
Check if an expression is a function call
}
