% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{isolate_capture}
\alias{isolate_capture}
\title{Isolate code by realizing inputs}
\usage{
isolate_capture(capture, envir = parent.frame())
}
\arguments{
\item{capture}{A `code_capture` object.}

\item{envir}{An environment in which to evaluate inputs (default: `parent.frame()`).}
}
\value{
A new `code_capture` object with input assignments prepended. Issues warnings for any inputs not found in the environment
}
\description{
Evaluates only the inputs used before they are defined, and prepends assignments
so the capture becomes self-contained.
}
