% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{length.code_capture}
\alias{length.code_capture}
\title{Length of a code_capture object}
\usage{
\method{length}{code_capture}(x)
}
\arguments{
\item{x}{`code_capture` object.}
}
\value{
The number of expressions in the capture
}
\description{
Length of a code_capture object
}
