% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{new_callobj}
\alias{new_callobj}
\title{Create call object structure from an expression}
\usage{
new_callobj(expr, meta = list())
}
\arguments{
\item{expr}{A call}

\item{meta}{A named list of metadata}
}
\value{
A list with expr and meta fields
}
\description{
Create call object structure from an expression
}
\keyword{internal}
