% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_calls}
\alias{normalize_calls}
\title{Normalize all calls inside an expression}
\usage{
normalize_calls(expr, env = parent.frame(), partial_match = TRUE)
}
\arguments{
\item{expr}{A call, expression, or list of calls.}

\item{env}{Environment to find functions (default: `parent.frame()`).}

\item{partial_match}{Logical; whether to allow partial matching of argument names (default: TRUE).}
}
\value{
A normalized version of the expression or block.
}
\description{
Recursively normalizes a call or list of calls by naming unnamed arguments
where safe and possible.
}
