% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{+.code_capture}
\alias{+.code_capture}
\title{Concatenate two code_capture objects}
\usage{
\method{+}{code_capture}(e1, e2)
}
\arguments{
\item{e1}{First `code_capture` object.}

\item{e2}{Second `code_capture` object.}
}
\value{
A new combined `code_capture` object.
}
\description{
Adds two `code_capture` objects together using `+`, combining expressions and metadata.
}
