% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{print.code_capture}
\alias{print.code_capture}
\title{Print method for code_capture objects}
\usage{
\method{print}{code_capture}(x, with_meta = TRUE, ...)
}
\arguments{
\item{x}{A `code_capture` object}

\item{with_meta}{Boolean; Should metadata be printed}

\item{...}{Ignored}
}
\description{
Print method for code_capture objects
}
