% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{print.code_diagnosis}
\alias{print.code_diagnosis}
\title{Print method for code diagnosis}
\usage{
\method{print}{code_diagnosis}(x, ...)
}
\arguments{
\item{x}{An object of class `"code_diagnosis"`.}

\item{...}{Ignored.}
}
\value{
Invisibly returns the diagnosis object.
}
\description{
Displays the results of a `code_diagnosis` object in a readable format.
}
