% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{process_capture}
\alias{process_capture}
\title{Realize and/or evaluate a capture}
\usage{
process_capture(
  capture,
  realize = FALSE,
  evaluate = FALSE,
  envir = parent.frame()
)
}
\arguments{
\item{capture}{A `code_capture` object}

\item{realize}{Logical; whether to realize arguments}

\item{evaluate}{Logical; whether to evaluate expressions}

\item{envir}{Environment to use}
}
\value{
A processed `code_capture` object
}
\description{
Realize and/or evaluate a capture
}
\keyword{internal}
