% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{process_expressions}
\alias{process_expressions}
\title{Process expressions into `callobj` list}
\usage{
process_expressions(expr_list)
}
\arguments{
\item{expr_list}{A list of unevaluated calls}
}
\value{
A list of capture_call structures
}
\description{
Process expressions into `callobj` list
}
\keyword{internal}
