% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{remove_arg}
\alias{remove_arg}
\title{Remove an argument from a call}
\usage{
remove_arg(expr, name)
}
\arguments{
\item{expr}{A call or list of calls.}

\item{name}{The argument name to remove.}
}
\value{
The modified call (or list of calls).
}
\description{
Deletes a named argument from a call expression.
}
