% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R, R/manipulate.R
\name{scan_expr_tree}
\alias{scan_expr_tree}
\title{Recursively scan an expression tree for matching nodes}
\usage{
scan_expr_tree(expr, predicate)

scan_expr_tree(expr, predicate)
}
\arguments{
\item{expr}{A call, symbol, or expression object.}

\item{predicate}{A function taking a node and returning TRUE/FALSE.}
}
\value{
A list of matching sub-expressions.

A list of matching sub-expressions.
}
\description{
Traverses an expression tree and collects nodes where the predicate returns TRUE.

Traverses an expression tree and collects nodes where the predicate returns TRUE.
}
\keyword{internal}
