% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_all_metadata}
\alias{set_all_metadata}
\title{Replace all metadata in a code_capture object}
\usage{
set_all_metadata(capture, meta)
}
\arguments{
\item{capture}{A `code_capture` object.}

\item{meta}{A list of named lists containing metadata.}
}
\value{
A modified `code_capture` object.
}
\description{
Updates the metadata for a specific expression in a `code_capture` object.
}
