% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{set_arg}
\alias{set_arg}
\title{Set or add an argument in a call}
\usage{
set_arg(expr, name, value)
}
\arguments{
\item{expr}{A expression or callobj.}

\item{name}{The argument name.}

\item{value}{The new value for the argument.}
}
\value{
The modified call (or list of calls).
}
\description{
Sets an argument to a new value in a call expression (or adds it if not present).
}
