% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_expressions}
\alias{set_expressions}
\title{Replace all expressions in a code_capture object}
\usage{
set_expressions(capture, exprs)
}
\arguments{
\item{capture}{A `code_capture` object.}

\item{exprs}{A list of call objects}
}
\value{
A modified `code_capture` object.
}
\description{
Replace all expressions in a code_capture object
}
