% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{set_function}
\alias{set_function}
\title{Set the function of a call expression}
\usage{
set_function(expr, fn)
}
\arguments{
\item{expr}{A call or callobj.}

\item{fn}{A symbol, call, or character to set as the new function.}
}
\value{
The modified call or callobj (same type as input).
}
\description{
Replaces the function being called.
}
