% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_metadata}
\alias{set_metadata}
\title{Replace metadata in a code_capture object}
\usage{
set_metadata(capture, i, meta)
}
\arguments{
\item{capture}{A `code_capture` object.}

\item{i}{Integer index of the metadata entry to replace.}

\item{meta}{A named list containing metadata.}
}
\value{
A modified `code_capture` object.
}
\description{
Updates the metadata for a specific expression in a `code_capture` object.
}
