% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{set_rhs}
\alias{set_rhs}
\title{Set the right-hand side of an assignment}
\usage{
set_rhs(expr, rhs)
}
\arguments{
\item{expr}{A call or callobj.}

\item{rhs}{An expression for the new RHS.}
}
\value{
The modified call or callobj (same type as input).
}
\description{
Replaces the RHS value or expression in an assignment call.
}
