% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{sort_capture}
\alias{sort_capture}
\title{Sort expressions in a code_capture object}
\usage{
sort_capture(capture)
}
\arguments{
\item{capture}{A `code_capture` object.}
}
\value{
A new `code_capture` object with sorted expressions.
}
\description{
Sorts constant assignments and function definitions to the top,
preserving relative order and associated metadata.
}
