% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{standardize_assignments}
\alias{standardize_assignments}
\title{Standardize assignment operators to `<-`}
\usage{
standardize_assignments(capture)
}
\arguments{
\item{capture}{A `code_capture` object.}
}
\value{
A modified `code_capture` object.
}
\description{
Rewrites any `=` assignments to use `<-` for consistency.
}
