% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{[.callobj}
\alias{[.callobj}
\title{Extract an argument by name or position}
\usage{
\method{[}{callobj}(x, i, ...)
}
\arguments{
\item{x}{A call object.}

\item{i}{A character (name) or numeric (position).}

\item{...}{Ignored. Included for method consistency.}
}
\description{
Extract an argument by name or position
}
