% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{[.capture_object}
\alias{[.capture_object}
\title{Subset a capture_object}
\usage{
\method{[}{capture_object}(x, i, ...)
}
\arguments{
\item{x}{A `capture_object`.}

\item{i}{Subset index.}

\item{...}{Additional arguments (ignored).}
}
\value{
A subsetted `capture_object`.
}
\description{
Provides subsetting (`[`) for `capture_object` while preserving the class.
}
