% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{[.code_capture}
\alias{[.code_capture}
\title{Subset a code_capture object}
\usage{
\method{[}{code_capture}(x, i, ...)
}
\arguments{
\item{x}{A `code_capture` object.}

\item{i}{Subset indices.}

\item{...}{Ignored.}
}
\value{
A subsetted `code_capture` object.
}
\description{
Provides subsetting (`[`) for `code_capture` while preserving the class and metadata.
}
