% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{[[.code_capture}
\alias{[[.code_capture}
\title{Extract an expression from a code_capture object}
\usage{
\method{[[}{code_capture}(x, i, ...)
}
\arguments{
\item{x}{A `code_capture` object.}

\item{i}{A single index.}

\item{...}{Ignored.}
}
\value{
A single expression (call object).
}
\description{
Allows using `[[` to access a specific captured expression.
}
