% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{[[<-.code_capture}
\alias{[[<-.code_capture}
\title{Replace an expression inside a code_capture object}
\usage{
\method{[[}{code_capture}(x, i) <- value
}
\arguments{
\item{x}{A `code_capture` object.}

\item{i}{Index to replace.}

\item{value}{A call object (expression) to assign.}
}
\value{
The modified `code_capture` object.
}
\description{
Allows using `[[<-` to assign a new expression at a specific index.
}
