% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{unwrap_expr}
\alias{unwrap_expr}
\title{Unwrap the outer function call of an expression}
\usage{
unwrap_expr(expr)
}
\arguments{
\item{expr}{A expression or callobj}
}
\value{
The unwrapped expression(s).
}
\description{
Removes the top-level function call and returns the first argument inside.
}
