% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{wrap_expr}
\alias{wrap_expr}
\title{Wrap an expression inside a function}
\usage{
wrap_expr(expr, wrapper_fn)
}
\arguments{
\item{expr}{A call or list of calls.}

\item{wrapper_fn}{A function name (symbol or character) to wrap with.}
}
\value{
A wrapped call (or list of calls).
}
\description{
Creates a new call by wrapping the given expression inside another function call.
}
