% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{capture}
\alias{capture}
\title{General capture interface}
\usage{
capture(x, script = NULL, envir = parent.frame())
}
\arguments{
\item{x}{A code block `{}` or missing}

\item{script}{Optional path to script}

\item{envir}{Environment to associate capture with}
}
\value{
A `code_capture` object or TRUE if session started
}
\description{
General capture interface
}
