% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{capture_script}
\alias{capture_script}
\title{Capture a script file}
\usage{
capture_script(
  path,
  encoding = "UTF-8",
  realize = FALSE,
  evaluate = FALSE,
  envir = parent.frame()
)
}
\arguments{
\item{path}{Path to script}

\item{encoding}{File encoding (default UTF-8)}

\item{realize}{Realize arguments (default FALSE)}

\item{evaluate}{Evaluate immediately (default FALSE)}

\item{envir}{Evaluation environment}
}
\value{
A `code_capture` object
}
\description{
Capture a script file
}
