% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{-.callobj}
\alias{-.callobj}
\title{Remove arguments from a call}
\usage{
\method{-}{callobj}(x, y)
}
\arguments{
\item{x}{A call object.}

\item{y}{A character vector (names) or numeric vector (positions).}
}
\description{
Remove arguments from a call
}
