% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{end_capture}
\alias{end_capture}
\title{Finalize session capture}
\usage{
end_capture(realize = FALSE, envir = parent.frame())
}
\arguments{
\item{realize}{Realize arguments (default FALSE)}

\item{envir}{Evaluation environment}
}
\value{
A list of `callobj` objects
}
\description{
Finalize session capture
}
