% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R, R/utils.R
\name{get_arguments}
\alias{get_arguments}
\title{Get the arguments of a call}
\usage{
get_arguments(expr)

get_arguments(expr)
}
\arguments{
\item{expr}{A call object.}
}
\value{
A list of argument expressions.

A list of arguments, or NULL if not a call.
}
\description{
Returns all arguments of a call, excluding the function.

Returns a list of arguments from a call expression.
}
