% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{+.callobj}
\alias{+.callobj}
\title{Add or modify arguments in a call}
\usage{
\method{+}{callobj}(e1, e2)
}
\arguments{
\item{e1}{A call object.}

\item{e2}{A named list (for setting arguments) or a single value (for appending).}
}
\description{
Add or modify arguments in a call
}
