% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{start_capture}
\alias{start_capture}
\title{Start capturing top-level expressions}
\usage{
start_capture(clear = TRUE, envir = parent.frame(), nframe = sys.nframe())
}
\arguments{
\item{clear}{Whether to clear previous buffer}

\item{envir}{Environment to associate capture with}

\item{nframe}{Internal safeguard}
}
\value{
Invisibly TRUE
}
\description{
Start capturing top-level expressions
}
